requirements.txt
tqdm
torchaudio
torchcodec
pandas
datasets==3.6.0
pyannote.audio
librosa
soundfile
transformers